+++ packages/boring-table-core/src/utils/index.ts
@@ -1,4 +1,4 @@
 export * from './createBoringTable';
 export * from './createOptions';
-export * from './debug';
+export * from './logger';